parser/htmlparser/tests/mochitest/test_bug396568.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=396568
-->
<head>
  <title>Test for Bug 396568</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <!--These script tags are part of the test, complete with indentation!-->
  <script language="javascript"






       
          type="text/javascript"></script>
  <script language="javascript"
          type="text/javascript"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=396568">Mozilla Bug 396568</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 396568 **/

 var reached = false;
 try {
   noSuchIdentifier;
   reached = true;
 } catch (e) {
   is(e.lineNumber, 36, "Unexpected line number"); //this line number is dependent on the line number of noSuchIdentifier
 }
 is(reached, false, "Exception needed to be thrown");
 
</script>
</pre>
</body>
</html>
